pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "overnight-predict"
version = "1.0.0"
description = "AI-powered autonomous code generation with question prediction and self-correcting accuracy mechanisms"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "OvernightPredict Team"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "aiohttp>=3.10.0",
    "asyncio-throttle>=1.0.0",
    "redis>=5.2.0",
    "aiosqlite>=0.20.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "structlog>=24.4.0",
    "rich>=13.9.0",
    "typer>=0.12.0",
    "httpx>=0.27.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.5.0",
    "sentence-transformers>=3.2.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "watchdog>=5.0.0",
    "aiofiles>=24.1.0",
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "mypy>=1.11.0",
    "ruff>=0.6.0",
    "black>=24.8.0",
    "pre-commit>=3.8.0",
]

[project.scripts]
overnight = "src.cli:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "ANN", "B", "C4", "SIM"]
ignore = ["ANN101", "ANN102", "ANN401"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
